Update all the things
